refactor: simplify modal height management and improve responsiveness
