+++ src/constants/index.ts
@@ -1,16 +1,4 @@
-import { ModalHeights, Routes } from '../types';
 import { IAppearance } from '../types';
-
-export const MODAL_HEIGHTS: ModalHeights = {
-  [Routes.PROFILE]: 473,
-  [Routes.WAITING]: 340,
-  [Routes.SIGN_TRANSACTION]: 458.4,
-  [Routes.SUCCESSFUL]: 328,
-  [Routes.ONBOARDING]: 520,
-  [Routes.SEND]: 326,
-  [Routes.ACTIVITY]: 408,
-  [Routes.OTP]: 424,
-};
 
 export const HORIZON_SERVERS = {
   public: 'https://horizon.stellar.org',
